+++ test/defaultSkin/array.react.test.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { mount } from 'enzyme'
-import { act } from 'react-dom/test-utils'
 
 import config from '../utils/enzymeConfig'
 import {