+++ test/utils/changeField.js
@@ -2,7 +2,7 @@
 // that doesn't seem to be in the simulated input.instance()
 
 
-export function changeInput(input, value) {
+export function changeInput(input, value, attr = 'value') {
   // FIXME I don't know why I do this, but it looks
   // like input.simulate() is not updating dom as a
   // browser would do or there's something I don't
@@ -12,6 +12,7 @@ export function changeInput(input, value) {
   return input.simulate('change', {
     target: {
       ...input.props(),
+      [attr]: value,
       value
     }
   })
@@ -19,4 +20,5 @@ export function changeInput(input, value) {
 
 export const changeSelect = changeInput
 
-export const changeCheckbox = changeInput
+export const changeCheckbox = (input, value) =>
+  changeInput(input, value, 'checked')